requirements.txt
anyio==4.1.0
appnope==0.1.3
argcomplete @ file:///private/tmp/python-argcomplete-20231211-42231-mc1bv2/argcomplete-3.2.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
boto3==1.34.0
botocore==1.34.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
docutils==0.20.1
environ==1.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.12.4
fqdn==1.5.1
ga4gh-cli @ file:///Users/pavelnikonorov/projects/ga4gh-cli/dist/ga4gh-cli-0.1.6.tar.gz#sha256=b12ff4b0037674dc06e7d3275cd012acd76d89b763bfe7b3fcf61c6da6f55135
ghp-import==2.1.0
idna==3.6
importlib-metadata==7.0.0
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==24.3.0
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.0.2
mkdocs==1.4.2
mkdocs-material==9.0.12
mkdocs-material-extensions==1.1.1
mkdocs-video==1.4.0
more-itertools==10.1.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
nh3==0.2.15
notebook_shim==0.2.3
overrides==7.4.0
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pymdown-extensions==9.9.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==25.1.2
readme-renderer==42.0
referencing==0.32.0
regex==2022.10.31
requests==2.26.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.13.2
s3transfer==0.9.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
terminado==0.18.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.0
twine==4.0.2
types-python-dateutil==2.8.19.14
uri-template==1.3.0
urllib3==1.26.18
virtualenv==20.24.5
watchdog==2.2.1
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
zipp==3.17.0
